move ready functions to server.cpp, handle unready when players leave
