+++ game_patch/multi/server_internal.h
@@ -104,14 +104,13 @@ void cleanup_win32_server_console();
 void handle_vote_command(std::string_view vote_name, std::string_view vote_arg, rf::Player* sender);
 std::set<rf::Player*> get_current_player_list(bool include_browsers);
 std::pair<bool, std::string> is_level_name_valid(std::string_view level_name_input);
-void handle_ready_command(rf::Player* sender);
-void handle_unready_command(rf::Player* sender);
 bool is_player_in_match(rf::Player* player);
 bool is_player_ready(rf::Player* player);
 void update_pre_match_powerups(rf::Player* player);
 void start_match();
 void cancel_match();
 void start_pre_match();
+void set_ready_status(rf::Player* player, bool is_ready);
 void server_vote_do_frame();
 void init_server_commands();
 void extend_round_time(int minutes);
